requirements/base.txt
# Django Core
Django>=5.0,<5.1
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0,<25.0

# Database
psycopg2-binary>=2.9,<3.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0

# Media & Storage
Pillow>=10.0,<11.0
django-storages>=1.14,<2.0
boto3>=1.34,<2.0

# Caching
django-redis>=5.4,<6.0

# Environment
python-decouple>=3.8,<4.0

# API Documentation
drf-spectacular>=0.27,<1.0

# Utils
python-slugify>=8.0,<9.0
requests>=2.31,<3.0

# YouTube thumbnail extraction
yt-dlp>=2024.1,<2025.0
